feat: Fix Django QuerySet TypeError in AnalyticsViewSet
This commit resolves an Internal Server Error that was being caused by an incorrect Django ORM query in the open_shifts_by_branch method.

The TypeError: QuerySet.annotate() received non-expression(s): branch__name was occurring because the alues() method was being passed a keyword argument (
ame='branch__name') that was not an expression.

To fix this, the query has been refactored to use Django's F() expression to properly alias the ranch__name field to
ame. This ensures the query is correctly interpreted by the Django ORM, eliminating the TypeError and allowing the endpoint to return the expected data.